% Auto-generated: do not edit by hand
\name{plotlyDashKetcher}

\alias{plotlyDashKetcher}

\title{PlotlyDashKetcher component}

\description{

}

\usage{
plotlyDashKetcher(id=NULL, buttonLabel=NULL, molecule=NULL)
}

\arguments{
\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{buttonLabel}{Character. }

\item{molecule}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

